fix(auth): github Provider not setup
Since github provider was not setup in auth.ts i.e auth.js config file, the user was being redirected to auth.js default auth/signup landing page. Since only Google auth was setup, we see a button for Google auth there instead.

The fix is to setup Github provider in auth.ts line 30.

Also, this has shed light on another issue with the login system: GitHub apps can only access emails that are public. If the user's email is not public, we won't get it, and since it has the @unique property in the schema, it causes an error in user creation.

Closes #1